+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove extra whitespace in posts without text.
 - Remove extra whitespace around "Publish" button.
 - Move preview button to the right of character counter.
+- Display character counter only when less than 100 characters left.
 
 ### Fixed